Remove separate TypeGPU folder in the examples list
