+++ apps/paper/src/App.tsx
@@ -26,7 +26,7 @@ import { DeferedRendering, ShadowMapping } from "./ShadowMapping";
 import { SamplerParameters } from "./Sampler";
 import { ReversedZ } from "./ReversedZ";
 import { ThreeJS } from "./ThreeJS";
-import { TypeGPU } from "./TypeGPU";
+import { GradientTiles } from "./GradientTiles";
 
 const Stack = createNativeStackNavigator<Routes>();
 
@@ -70,7 +70,7 @@ function App() {
           <Stack.Screen name="Tests">
             {(props) => <Tests {...props} assets={assets} />}
           </Stack.Screen>
-          <Stack.Screen name="TypeGPU" component={TypeGPU} />
+          <Stack.Screen name="GradientTiles" component={GradientTiles} />
         </Stack.Navigator>
       </NavigationContainer>
     </GestureHandlerRootView>
