+++ apps/paper/src/GradientTiles/gradientWgsl.ts
