+++ apps/paper/src/Route.ts
@@ -20,5 +20,5 @@ export type Routes = {
   Wireframe: undefined;
   Resize: undefined;
   Tests: undefined;
-  TypeGPU: undefined;
+  GradientTiles: undefined;
 };
