--- apps/paper/src/TypeGPU/Routes.ts
@@ -1,4 +1,0 @@
-export type Routes = {
-  List: undefined;
-  GradientTiles: undefined;
-};
