--- apps/paper/src/TypeGPU/index.tsx
@@ -1,37 +1,0 @@
-import React, { useEffect } from "react";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { warnIfNotHardwareAccelerated } from "react-native-wgpu";
-
-import { GradientTiles } from "./GradientTiles";
-import type { Routes } from "./Routes";
-import { List } from "./List";
-
-const Stack = createNativeStackNavigator<Routes>();
-export const TypeGPU = () => {
-  useEffect(() => {
-    navigator.gpu.requestAdapter().then((adapter) => {
-      if (adapter) {
-        warnIfNotHardwareAccelerated(adapter);
-      }
-    });
-  }, []);
-  return (
-    <Stack.Navigator>
-      <Stack.Screen
-        name="List"
-        component={List}
-        options={{
-          title: "💜 TypeGPU",
-          header: () => null,
-        }}
-      />
-      <Stack.Screen
-        name="GradientTiles"
-        component={GradientTiles}
-        options={{
-          title: "🌈 Gradient Tiles",
-        }}
-      />
-    </Stack.Navigator>
-  );
-};
